Model Components

Polynomial Regression

The PolynomialRegression model fits a polynomial of a specified degree to the time series data.

TemporalMixtureModels.PolynomialRegression — Method

Create a polynomial regression mixture model component of specified degree.

Arguments

degree::Int: Degree of the polynomial.

Example

Creating a polynomial regression model of degree 2:

model = PolynomialRegression(2)

Ridge Regression

The Ridge model fits a polynomial regression with L2 regularization (ridge regression)

TemporalMixtureModels.RidgePolynomialRegression — Method

Create a ridge (L2) polynomial regression mixture model component of specified degree and regularization parameter.

Arguments

degree::Int: Degree of the polynomial.

lambda::T: Regularization parameter (L2 penalty).

Example

Creating a ridge polynomial regression model of degree 2 with lambda = 0.1:

model = RidgePolynomialRegression(2, 0.1)

Lasso Regression

The Lasso model fits a polynomial regression with L1 regularization (lasso regression)

TemporalMixtureModels.LassoPolynomialRegression — Method

Create a lasso (L1) polynomial regression mixture model component of specified degree and regularization parameter.

Arguments

degree::Int: Degree of the polynomial.

lambda::T: Regularization parameter (L1 penalty).

Example

Creating a lasso polynomial regression model of degree 2 with lambda = 0.1:

model = LassoPolynomialRegression(2, 0.1)

Custom Models

Custom model components can be implemented by subtyping the AbstractMixtureModelComponent and implementing the required methods. This allows for flexibility in defining new model types and behaviors tailored to specific use cases. See the tutorial on Implementing Custom Components for how to create your own model components.